+++ wazo_webhookd/plugins/api/http.py
@@ -1,10 +1,11 @@
 # Copyright 2017-2023 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
+from __future__ import annotations
 
 import yaml
 import logging
 
-from flask import make_response
+from flask import make_response, Response
 from flask_restful import Resource
 from xivo.chain_map import ChainMap
 from xivo.http_helpers import reverse_proxy_fix_api_spec
@@ -16,7 +17,7 @@ logger = logging.getLogger(__name__)
 class SwaggerResource(Resource):
     api_filename = "api.yml"
 
-    def get(self):
+    def get(self) -> tuple[dict[str, str], int] | Response:
         api_spec = ChainMap(
             *load_all_api_specs('wazo_webhookd.plugins', self.api_filename)
         )