+++ wazo_webhookd/plugins/config/http.py
@@ -1,5 +1,8 @@
-# Copyright 2017-2021 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2023 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
 
 from flask import request
 from wazo_webhookd.rest_api import AuthResource
@@ -9,19 +12,23 @@ from jsonpatch import JsonPatch
 
 from .schemas import config_patch_schema
 
+if TYPE_CHECKING:
+    from .service import ConfigService
+    from ...types import WebhookdConfigDict
+
 
 class ConfigResource(AuthResource):
-    def __init__(self, config_service):
+    def __init__(self, config_service: ConfigService) -> None:
         self._config_service = config_service
 
     @required_master_tenant()
     @required_acl('webhookd.config.read')
-    def get(self):
+    def get(self) -> tuple[WebhookdConfigDict, int]:
         return self._config_service.get_config(), 200
 
     @required_master_tenant()
     @required_acl('webhookd.config.update')
-    def patch(self):
+    def patch(self) -> tuple[WebhookdConfigDict, int]:
         config_patch = config_patch_schema.load(request.get_json(), many=True)
         config = self._config_service.get_config()
         patched_config = JsonPatch(config_patch).apply(config)