+++ wazo_webhookd/plugins/config/plugin.py
@@ -1,12 +1,18 @@
-# Copyright 2017-2020 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2023 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
 
 from .http import ConfigResource
 from .service import ConfigService
 
+if TYPE_CHECKING:
+    from ...types import PluginDependencyDict
+
 
 class Plugin:
-    def load(self, dependencies):
+    def load(self, dependencies: PluginDependencyDict) -> None:
         api = dependencies['api']
         config = dependencies['config']
         config_service = ConfigService(config)