+++ wazo_webhookd/plugins/mobile/plugin.py
@@ -1,16 +1,22 @@
-# Copyright 2015-2023 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2023 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+from .http import NotificationResource
+
 if TYPE_CHECKING:
     from ...types import PluginDependencyDict
-
-from .http import SwaggerResource
 
 
 class Plugin:
     def load(self, dependencies: PluginDependencyDict) -> None:
         api = dependencies['api']
-        api.add_resource(SwaggerResource, '/api/api.yml')
+        config = dependencies['config']
+
+        api.add_resource(
+            NotificationResource,
+            '/mobile/notifications',
+            resource_class_args=[config],
+        )