+++ wazo_webhookd/plugins/subscription/bus.py
@@ -159,15 +159,14 @@ class SubscriptionBusEventHandler:
             if extras := entry_point.extras:
                 entry_point_name += f' [{",".join(extras)}]'
 
-            hook_runner_task.s(
+            task_args = (
                 hook_uuid,
                 entry_point_name,
-                # self._config is a custom ChainMap and not a dict so it has this extra property,
-                # however that was not type-able
-                self._config.data,  # type: ignore
+                dict(self._config),
                 subscription,
                 payload,
-            ).apply_async()
+            )
+            hook_runner_task.delay(*task_args)
         except kombu.exceptions.OperationalError:
             # NOTE(sileht): That's not perfect in real life, because if celery
             # lose the connection, we have a good chance that our bus lose it